+++ cmd/kubeadm/app/cmd/phases/join/checketcd.go
