+++ cmd/kubeadm/app/cmd/phases/join/controlplane.go
@@ -30,75 +30,12 @@ import (
 	certsphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/certs"
 	"k8s.io/kubernetes/cmd/kubeadm/app/phases/controlplane"
 	kubeconfigphase "k8s.io/kubernetes/cmd/kubeadm/app/phases/kubeconfig"
-	"k8s.io/kubernetes/pkg/util/normalizer"
 )
-
-var (
-	controlPlaneExample = normalizer.Examples(`
-		# Generates all static Pod manifest files for control plane components,
-		# functionally equivalent to what is generated by kubeadm init.
-		kubeadm init phase control-plane all
-
-		# Generates all static Pod manifest files using options read from a configuration file.
-		kubeadm init phase control-plane all --config config.yaml
-		`)
-
-	controlPlanePhaseProperties = map[string]struct {
-		name  string
-		short string
-	}{
-		kubeadmconstants.KubeAPIServer: {
-			name:  "apiserver",
-			short: getPhaseDescription(kubeadmconstants.KubeAPIServer),
-		},
-		kubeadmconstants.KubeControllerManager: {
-			name:  "controller-manager",
-			short: getPhaseDescription(kubeadmconstants.KubeControllerManager),
-		},
-		kubeadmconstants.KubeScheduler: {
-			name:  "scheduler",
-			short: getPhaseDescription(kubeadmconstants.KubeScheduler),
-		},
-	}
-)
-
-type controlPlaneData interface {
-	Cfg() *kubeadmapi.InitConfiguration
-	KubeConfigDir() string
-	ManifestDir() string
-}
 
 type controlPlanePrepareData interface {
 	Cfg() *kubeadmapi.JoinConfiguration
 	ClientSetFromFile(string) (*clientset.Clientset, error)
 	InitCfg() (*kubeadmapi.InitConfiguration, error)
-}
-
-func getPhaseDescription(component string) string {
-	return fmt.Sprintf("Generates the %s static Pod manifest", component)
-}
-
-// NewControlPlanePhase creates a kubeadm workflow phase that implements bootstrapping the control plane.
-func NewControlPlanePhase() workflow.Phase {
-	phase := workflow.Phase{
-		Name:  "control-plane",
-		Short: "Generates all static Pod manifest files necessary to establish the control plane",
-		Long:  cmdutil.MacroCommandLongDescription,
-		Phases: []workflow.Phase{
-			{
-				Name:           "all",
-				Short:          "Generates all static Pod manifest files",
-				InheritFlags:   getControlPlanePhaseFlags("all"),
-				Example:        controlPlaneExample,
-				RunAllSiblings: true,
-			},
-			newControlPlaneSubphase(kubeadmconstants.KubeAPIServer),
-			newControlPlaneSubphase(kubeadmconstants.KubeControllerManager),
-			newControlPlaneSubphase(kubeadmconstants.KubeScheduler),
-		},
-		Run: runControlPlanePhase,
-	}
-	return phase
 }
 
 // NewControlPlanePreparePhase creates a kubeadm workflow phase that implements the preparation of the node to serve a control plane
@@ -108,51 +45,17 @@ func NewControlPlanePreparePhase() workflow.Phase {
 		Short: "Prepares the machine for serving a control plane.",
 		Long:  cmdutil.MacroCommandLongDescription,
 		Phases: []workflow.Phase{
+			{
+				Name:           "all",
+				Short:          "Prepares the machine for serving a control plane.",
+				InheritFlags:   getControlPlanePreparePhaseFlags(),
+				RunAllSiblings: true,
+			},
 			newControlPlanePrepareCertsSubphase(),
 			newControlPlanePrepareKubeconfigSubphase(),
 			newControlPlanePrepareManifestsSubphases(),
 		},
 	}
-}
-
-func newControlPlaneSubphase(component string) workflow.Phase {
-	phase := workflow.Phase{
-		Name:         controlPlanePhaseProperties[component].name,
-		Short:        controlPlanePhaseProperties[component].short,
-		Run:          runControlPlaneSubphase(component),
-		InheritFlags: getControlPlanePhaseFlags(component),
-	}
-	return phase
-}
-
-func getControlPlanePhaseFlags(name string) []string {
-	flags := []string{
-		options.CfgPath,
-		options.CertificatesDir,
-		options.KubernetesVersion,
-		options.ImageRepository,
-	}
-	if name == "all" || name == kubeadmconstants.KubeAPIServer {
-		flags = append(flags,
-			options.APIServerAdvertiseAddress,
-			options.APIServerBindPort,
-			options.APIServerExtraArgs,
-			options.FeatureGatesString,
-			options.NetworkingServiceSubnet,
-		)
-	}
-	if name == "all" || name == kubeadmconstants.KubeControllerManager {
-		flags = append(flags,
-			options.ControllerManagerExtraArgs,
-			options.NetworkingPodSubnet,
-		)
-	}
-	if name == "all" || name == kubeadmconstants.KubeScheduler {
-		flags = append(flags,
-			options.SchedulerExtraArgs,
-		)
-	}
-	return flags
 }
 
 func getControlPlanePreparePhaseFlags() []string {
@@ -173,7 +76,7 @@ func newControlPlanePrepareCertsSubphase() workflow.Phase {
 		Name:         "certs",
 		Short:        "Generates the certificates for the new control plane components",
 		Run:          runControlPlanePrepareCertsPhaseLocal,
-		InheritFlags: getControlPlanePreparePhaseFlags(),
+		InheritFlags: getControlPlanePreparePhaseFlags(), //NB. eventually in future we would like to break down this in sub phases for each cert or add the --csr option
 	}
 }
 
@@ -182,59 +85,37 @@ func newControlPlanePrepareKubeconfigSubphase() workflow.Phase {
 		Name:         "kubeconfig",
 		Short:        "Generates the kubeconfig for the new control plane components",
 		Run:          runControlPlanePrepareKubeconfigPhaseLocal,
-		InheritFlags: getControlPlanePreparePhaseFlags(),
+		InheritFlags: getControlPlanePreparePhaseFlags(), //NB. eventually in future we would like to break down this in sub phases for each kubeconfig
 	}
 }
 
 func newControlPlanePrepareManifestsSubphases() workflow.Phase {
 	return workflow.Phase{
-		Name:  "manifests",
-		Short: "Generates the manifests for the new control plane components",
-		Phases: []workflow.Phase{
-			{
-				Name:           "all",
-				Short:          "Generates all static Pod manifest files",
-				InheritFlags:   getControlPlanePreparePhaseFlags(),
-				RunAllSiblings: true,
-			},
-			newControlPlanePrepareSubphase(kubeadmconstants.KubeAPIServer),
-			newControlPlanePrepareSubphase(kubeadmconstants.KubeControllerManager),
-			newControlPlanePrepareSubphase(kubeadmconstants.KubeScheduler),
-		},
-		InheritFlags: getControlPlanePreparePhaseFlags(),
+		Name:         "manifests",
+		Short:        "Generates the manifests for the new control plane components",
+		Run:          runControlPlaneSubphase,
+		InheritFlags: getControlPlanePreparePhaseFlags(), //NB. eventually in future we would like to break down this in sub phases for each component
 	}
 }
 
-func newControlPlanePrepareSubphase(component string) workflow.Phase {
-	return workflow.Phase{
-		Name:         controlPlanePhaseProperties[component].name,
-		Short:        controlPlanePhaseProperties[component].short,
-		Run:          runControlPlanePrepareJoinSubphase(component),
-		InheritFlags: getControlPlanePreparePhaseFlags(),
-	}
-}
-
-func runControlPlanePhase(c workflow.RunData) error {
-	data, ok := c.(controlPlaneData)
+func runControlPlaneSubphase(c workflow.RunData) error {
+	data, ok := c.(controlPlanePrepareData)
 	if !ok {
-		return errors.New("control-plane phase invoked with an invalid data struct")
+		return errors.New("control-plane-prepare phase invoked with an invalid data struct")
 	}
 
-	fmt.Printf("[control-plane] Using manifest folder %q\n", data.ManifestDir())
-	return nil
-}
-
-func runControlPlaneSubphase(component string) func(c workflow.RunData) error {
-	return func(c workflow.RunData) error {
-		data, ok := c.(controlPlaneData)
-		if !ok {
-			return errors.New("control-plane phase invoked with an invalid data struct")
-		}
-		cfg := data.Cfg()
-
-		fmt.Printf("[control-plane] Creating static Pod manifest for %q\n", component)
-		return controlplane.CreateStaticPodFiles(data.ManifestDir(), &cfg.ClusterConfiguration, &cfg.LocalAPIEndpoint, component)
+	// Skip if this is not a control plane
+	if data.Cfg().ControlPlane == nil {
+		return nil
 	}
+
+	cfg, err := data.InitCfg()
+	if err != nil {
+		return err
+	}
+
+	// Generate missing certificates (if any)
+	return controlplane.CreateInitStaticPodManifestFiles(kubeadmconstants.GetStaticPodDirectory(), cfg)
 }
 
 func runControlPlanePrepareCertsPhaseLocal(c workflow.RunData) error {
